requirements.txt
Flask==2.3.3
httpx==0.27.0
Flask-Limiter==3.5.0
python-dotenv==1.0.1
